Updated the Room User dao to limit the amount of users that can be stored
