+++ app/src/main/java/com/receipttracker/ViewModel/common/RegisterViewModel.kt
@@ -1,0 +1,18 @@
+package com.receipttracker.ViewModel.common
+
+import com.receipttracker.model.User
+
+class RegisterViewModel : BaseViewModel<User>() {
+
+    override fun create(obj: User) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun update(obj: User) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun delete(obj: User) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+}
